requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
detect==2020.12.3
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
gitdb==4.0.11
grpcio==1.64.1
idna==3.7
intel-openmp==2021.4.0
ipython==8.26.0
itypes==1.2.0
jedi==0.19.1
kiwisolver==1.4.5
lapx==0.5.9.post1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocotools==2.0.8
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
stack-data==0.6.3
supervision==0.21.0rc5
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
tracker==0.1.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.48
ultralytics-copy==0.0.1.dev0
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
